sc2: Added a setting controlling if sc2 will boot in windowed mode
